+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - [**breaking**] Improve the action key parser to support multiple key bindings. Representation of the default action key is changed to `LAlt-S,LAlt-Enter` (<s>`LAlt-S/Enter`</s>). New syntax for the key bindings is: `<key1>,<key2>-<key3>-...-<key4>,<key5>`
 - [**breaking**] Rename `--keys` option to `--action-keys` including corresponding configuration file entries and environment variables.
+- Change default timeout value for area selection to 300 seconds
 - Upgrade dependencies in Cargo.toml
 
 ## [1.3.0] - 2021-03-20